Move getSerializableProperties helper to own file
