+++ src/testing/matchers/toEqualStrictTyped.ts
@@ -1,33 +1,7 @@
 import { iterableEquality } from "@jest/expect-utils";
 import type { MatcherFunction } from "expect";
 
-import { ApolloClient, ObservableQuery } from "@apollo/client/core";
-import { isPlainObject } from "@apollo/client/utilities";
-
-function isKnownClassInstance(value: unknown) {
-  return [ApolloClient, ObservableQuery].some((c) => value instanceof c);
-}
-
-function getSerializableProperties(obj: unknown): any {
-  if (Array.isArray(obj)) {
-    return obj.map((item) => getSerializableProperties(item));
-  }
-
-  if (isPlainObject(obj)) {
-    return Object.entries(obj).reduce(
-      (memo, [key, value]) => {
-        if (typeof value === "function" || isKnownClassInstance(value)) {
-          return memo;
-        }
-
-        return { ...memo, [key]: value };
-      },
-      {} as Record<string, any>
-    );
-  }
-
-  return obj;
-}
+import { getSerializableProperties } from "./utils/getSerializableProperties.js";
 
 export const toEqualStrictTyped: MatcherFunction<[value: any]> = function (
   actual,
